Externatlisation du contenu du modal dans un composant partagé
