+++ articles/voyage-v0.2.mdx
@@ -29,6 +29,8 @@ Voici ce qu'il y a de nouveau la version v0.2.
 
 Voici les perspectives pour la suite. 
 
+### Interface
+
 ### Destination 
 
 - quelques infos sur la destination, via Wikipedia
